# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Merchant registration",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Access and refresh tokens issued"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Merchant login",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh access token from httpOnly cookie",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout and invalidate refresh token",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/merchants/me": {
      "get": {
        "operationId": "MerchantsController_me",
        "summary": "Current merchant profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "merchants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "MerchantsController_update",
        "summary": "Update merchant profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "merchants"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/compliance/screen": {
      "post": {
        "operationId": "ComplianceController_screen",
        "summary": "OFAC screen a wallet address",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "compliance"
        ]
      }
    },
    "/invoices": {
      "post": {
        "operationId": "InvoicesController_create",
        "summary": "Create invoice",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "InvoicesController_list",
        "summary": "List invoices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invoices/public/{id}": {
      "get": {
        "operationId": "InvoicesController_publicInvoice",
        "summary": "Public invoice for payment page",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "invoices"
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "operationId": "InvoicesController_get",
        "summary": "Invoice detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/invoices/{id}/cancel": {
      "post": {
        "operationId": "InvoicesController_cancel",
        "summary": "Cancel pending invoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "invoices"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/payments/{id}/prepare-tx": {
      "get": {
        "operationId": "PaymentsController_prepare",
        "summary": "Build unsigned Stellar payment XDR",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payer",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{id}/stream": {
      "get": {
        "operationId": "PaymentsController_stream",
        "summary": "Real-time invoice status stream",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/webhooks": {
      "post": {
        "operationId": "WebhooksController_create",
        "summary": "Register webhook",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "WebhooksController_list",
        "summary": "List merchant webhooks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/webhooks/{id}": {
      "delete": {
        "operationId": "WebhooksController_remove",
        "summary": "Deactivate webhook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/webhooks/{id}/deliveries": {
      "get": {
        "operationId": "WebhooksController_deliveries",
        "summary": "Webhook delivery history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/webhooks/deliveries/{id}/retry": {
      "post": {
        "operationId": "WebhooksController_retry",
        "summary": "Retry failed delivery",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Stargate API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {}
  }
}